requirements.txt
-q -U google-genai
termcolor
python-dotenv
colorama
pyinstaller